Add lots more refactoring. Remove store, rename declarations. Get test basic working
