+++ src/core/core.ts
@@ -1,13 +1,13 @@
 import defaultsDeep from 'lodash/defaultsDeep';
 import kebabCase from 'lodash/kebabCase';
 import pluralize from 'pluralize';
-import { Config } from '../declarations/config.d';
-import { CustomRoute } from '../declarations/custom-routes.d';
-import { RequestData } from '../declarations/request.d';
-import { Route } from '../declarations/routes.d';
+import { Config } from '../typings/config.d';
+import { CustomRoute } from '../typings/custom-routes.d';
+import { RequestData } from '../typings/request.d';
+import { Route } from '../typings/routes.d';
 import { sanitizeUrl } from '../utils/url';
-import Defaults from './defaults';
 import Debugger from './../debug/debugger';
+import Defaults from './defaults';
 
 class Core {
   protected config: Config;
@@ -17,6 +17,7 @@ class Core {
   protected requestData: RequestData;
   protected routes: Config['routes'];
   protected urlParams: any[];
+  protected debugger: Debugger|null;
 
   constructor(config: Config) {
     this.config = Object.assign({}, defaultsDeep(config, Defaults));
@@ -31,11 +32,6 @@ class Core {
 
     this.initialize();
   }
-
-  /**
-   * Set any config overrides in this method when extending
-   */
-  boot() {}
 
   /**
    * The order of these are important.
@@ -62,57 +58,24 @@ class Core {
     this.initializeDebugger();
   }
 
-  /**
-   * Initialze the debugger if debug is set to true.
+   /**
+   * Set any config overrides in this method when extending
    */
-  private initializeDebugger(): void {
-    this.debugger = this.config.debug ? new Debugger(this) : false;
+  boot() {}
+
+  /**
+   * Sanitize the baseURL before sending it to the http service
+   */
+  private sanitizeBaseURL(): void {
+    this.config.baseURL = sanitizeUrl(this.config.baseURL, this.config.trailingSlash);
   }
 
   /**
-   * Initialize the API.
-   * consider making an adatper interface to talk to http methods
+   * Loop through the routes and set them
    */
-  private initializeHttp(): void {
-    const httpConfig = defaultsDeep(
-      {
-        baseURL: this.config.baseURL.replace(/\/$/, ''),
-      },
-      this.config.httpConfig,
-    );
-
-    const http = this.config.http;
-
-    this.http = new http(httpConfig);
-  }
-
-  /**
-   * Set up the custom routes if we have any
-   */
-  private defineCustomRoutes(): void {
-    // if we have custom routes, set up a name:route mapping
-    if (this.config.customRoutes.length) {
-      this.config.customRoutes.forEach((route: CustomRoute) => {
-        this.customRoutes[route.name] = route;
-      });
-    }
-  }
-
-  /**
-   * Resets the request data
-   */
-  protected resetRequestData(): void {
-    this.requestData = {
-      params: {},
-      options: {},
-    };
-  }
-
-  /**
-   * Reset an URL params set from a relationship
-   */
-  protected resetURLParams(): void {
-    this.urlParams = [];
+  private setRoutes(): void {
+    // ['model', 'collection'].forEach(route => this.setRoute(route));
+    [Route.MODEL, Route.COLLECTION].forEach(route => this.setRoute(route));
   }
 
   /**
@@ -142,23 +105,61 @@ class Core {
   }
 
   /**
-   * Loop through the routes and set them
+   * Set up the custom routes if we have any
    */
-  private setRoutes(): void {
-    [Route.MODEL, Route.COLLECTION].forEach(route => this.setRoute(route));
+  private defineCustomRoutes(): void {
+    // if we have custom routes, set up a name:route mapping
+    if (this.config.customRoutes.length) {
+      this.config.customRoutes.forEach((route: CustomRoute) => {
+        this.customRoutes[route.name] = route;
+      });
+    }
   }
 
   /**
-   * Sanitize the baseURL before sending it to the http service
+   * Initialize the API.
+   * consider making an adatper interface to talk to http methods
    */
-  private sanitizeBaseURL(): void {
-    this.config.baseURL = sanitizeUrl(this.config.baseURL, this.config.trailingSlash);
+  private initializeHttp(): void {
+    const httpConfig = defaultsDeep(
+      {
+        baseURL: this.config.baseURL.replace(/\/$/, ''),
+      },
+      this.config.httpConfig,
+    );
+
+    const http = this.config.http;
+
+    this.http = new http(httpConfig);
+  }
+
+  /**
+   * Initialze the debugger if debug is set to true.
+   */
+  private initializeDebugger(): void {
+    this.debugger = this.config.debug ? new Debugger(this) : null;
+  }
+
+  /**
+   * Resets the request data
+   */
+  protected resetRequestData(): void {
+    this.requestData = {
+      params: {},
+      options: {},
+    };
+  }
+
+  /**
+   * Reset an URL params set from a relationship
+   */
+  protected resetURLParams(): void {
+    this.urlParams = [];
   }
 
   /**
    * Getters
    */
-
   get collection() {
     this.currentRoute = Route.COLLECTION;
