--- src/declarations/custom-routes.d.ts
@@ -1,7 +1,0 @@
-import { RequestType } from './request';
-
-export declare interface CustomRoute {
-  name: string;
-  type: RequestType;
-  url: string;
-}
