--- src/declarations/routes.d.ts
@@ -1,5 +1,0 @@
-export declare enum Route {
-  ANY = 'any',
-  MODEL = 'model',
-  COLLECTION = 'collection',
-}
