+++ src/rapid.ts
@@ -4,3 +4,5 @@ import Auth from './auth';
 module.exports = Rapid;
 module.exports.Rapid = Rapid;
 module.exports.Auth = Auth;
+
+export default Rapid;
