--- test/store.test.ts
@@ -1,7 +1,0 @@
-import store from '../src/store/index';
-
-describe('The store actions and state work as they should', () => {
-  it('has currentRoute', () => {
-
-  });
-});
